refactor(point): policy, service 관련 파일들 service 하위 디렉터리로 이동
