+++ src/point/point.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { PointController } from './controller/point.controller';
-import { PointService } from './point.service';
+import { PointService } from './service/point.service';
 import { DatabaseModule } from 'src/database/database.module';
-import { PointPolicy } from './policy/point.policy';
+import { PointPolicy } from './service/policy/point.policy';
 import { PointRepository } from './repository/point.repository';
 
 @Module({
